Fix go build caching
I am not sure why, but Dockerfile-dev's pre-building of gocv.io/x/gocv,
which is intended to cache those objects so that they don't need
rebuilding again, was not working for me.  Fix this by mapping in a
cache directory from outside the container.

(Actually it's probably because I had to resync vendor/gocv.io/x/gocv
after originally building the go-dev-image...)