Minesweeper fixes
- Don't panic when we fail to spot any contours.  Instead just return
a tiny contour area, which will map through to negligible confidence.

- Computed distance to centre of target square was actually in cm, and
we need mm.  Multiply by 10 to fix that.

- More logging in a few places where we were missing useful numbers.

- After sitting on bomb, fix the transition back into
POSSIBLY_UNSAFE_FOR_SEARCH state.  Specifically, this should mean (1)
new target at least 400 away from walls, (2) reset approachingTarget
to false.  We weren't doing either of those things.

Which begs the question: how did it find a second bomb after Shaun
moved it?

The answer is that it will still work if the next bomb is already in
the bot's field of view.  Then it'll adjust heading as needed and work
its way towards it.  But in the more general case, we need the bot to
do a search rotation, and in order to do that it must be at a safe
distance from the walls.

The logs also showed "Sit on bomb!" many times before actually
arriving at the (presumed) actual second bomb position.  This was
caused by undercalculating the remaining distance.  So the bot kept
thinking it had arrived, prematurely.

Finally, when it really passed over the second bomb, we got the panic
because of not identifying any contours ahead.