Fix left/right mixup in lava palava
From the analysis of the first picture:
```
2024/04/19 13:18:06 >> RESULT: -0.007094 105.029431
```

First number is gradient and second number is whether the closest part
of the line is to the left (<100) or to the right (>100).  So in this
case negligible gradient and slightly to the right.  But then the
following code has sign errors and so the bot heads further to the
left instead of to the right.